requirements.txt
streamlit==0.84.0
scikit-learn==0.24.2
seaborn==0.11.1
plotly==5.1.0
pandas==1.3.0
numpy==1.19.5
matplotlib==3.4.2
Pillow==8.3.1